Create links carousel
