+++ package.json
@@ -18,6 +18,7 @@
     "@sanity/vision": "^3.44.0",
     "clsx": "^2.1.1",
     "graphql": "^16.8.1",
+    "keen-slider": "^6.8.6",
     "next": "14.2.3",
     "next-sanity": "^9.3.8",
     "react": "^18",
